Switch over to static schema generation.
Took a bit of doing.  DBIC::Schema::Loader will only make a schema using load_classes()
or load_namespaces() which spider all of lib/BackPAN/Index/, see non result classes and
freak out.  So we let it make a Schema class and throw it out.

We need to generate the database at build time so Schema::Loader can see it.