DBIx::Class::Schema::Loader is no longer needed to build the module.
Have to do it manually now rather than checking at build time.  It was causing
failures because it was still trying to load Schema::Loader before checking.
It could be changed to do the up to date check before loading modules, but it
doesn't seem worth the bother how often we change the schema.